Fix address of stack
No longer need to hardcode a 64KB stack, since the crt0 code adds the size
of the memory to 0x80000000. So if I return 0x800000 (128MB, the default in
QEMU) it will correctly start the stack at just below the end of the memory.

Note: `info mtree` in QEMU is very useful. It would be even more useful if
there was an option in QEMU to trap when accessing unmapped memory instead
of ignoring the write/returning 0