Add option for long-granularity swapping in quicksort with capabilities
To support capabilities, the fast case was extended from longs to capabilities.
This means that structs that were long-aligned but are not capability-aligned
fall through to the slow byte-by-byte case on 64-bit, as is the case for the
MiBench quicksort benchmark.
This commit restores long-granularity copying as an extra case in capability mode.
It also adds a redundat extra long-granularity case to non-cap-mode.
Perhaps an ideal compromise is to make the "middle" cases be 32-bit for both
capability and non-capability mode.